first pass refactor trace fetching
